requirements.txt
torch
diffusers
transformers
grpcio
grpcio-tools
Pillow
protobuf
pytest
accelerate
safetensors